+++ src/app/modules/demo/demo.component.ts
@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  templateUrl: './demo.component.html',
+  styleUrls: ['./demo.component.scss'],
 })
-export class AppComponent {
+export class DemoComponent {
   title = 'ng-tower-of-heroes-ts-app';
 }