+++ HISTORY.md
@@ -3,4 +3,5 @@
 1. [Installed `@angular/cli`: `ng`](https://angular.io/tutorial/first-app)
 1. [`ng new {my app}`](https://angular.io/cli)
 1. Installed linters and what/not
-1. [Replaced contents of `app.component.\* with "Hello World"](https://angular.io/tutorial/first-app/first-app-lesson-01)
+1. [`ng generate component "hello-world"`](https://angular.io/tutorial/first-app/first-app-lesson-01)
+1. Refactored `AppComponent` as `modules/demo/demo.component.ts`